### 2021-01-11 #2951 @whedon
[PRE REVIEW]: OpenRepGrid.ic: A software for Interpretive Clustering
**Submitting author:** @markheckmann (<a href="http://orcid.org/0000-0002-0736-7417">Mark Heckmann</a>)
**Repository:** <a href="https://github.com/markheckmann/OpenRepGrid.ic" target ="_blank">https://github.com/markheckmann/OpenRepGrid.ic</a>
**Version:** v0.5.0
