### 2021-01-14 #2951 @Kevin-Mattheus-Moerman
@markheckmann I've flagged this submission for additional scope review by our editorial board. In particular we will review if it is considered to have [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This is standard for submissions of this size (the core functionality appears to be captured by around 1661 lines of code).  This review will take about a week.

In the mean time, can you please review our documentation on [what the paper should contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), in particular the "Statement of need" section is missing. You can update the paper here by calling: `@whedon generate pdf`. Thanks.
