### 2021-01-19 #2951 @kthyng
Hi @markheckmann — the editorial board has determined that this submission is in scope, so we can proceed.

Please follow up on @Kevin-Mattheus-Moerman's comment above to add a Statement of Need section.

Also, add a sentence or two to the start of your paper so that a non-expert in your field can be oriented to where your software fits.
