+++ paper.md
@@ -43,7 +43,12 @@ pre {
   
 # Introduction
 
-The [`OpenRepGrid.ic` R package](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) is a browser-based software to perform *Interpretive Clustering* (IC) for [repertory grid](https://en.wikipedia.org/wiki/Repertory_grid) data. The software implements the IC method as described in detail in our recent publication [[@burr_qualitative_2020]](https://doi.org/10.1080/14780887.2020.1794088). While IC can also be conducted by hand, this is very time consuming, error-prone and only feasible for small-sized repertory grids. Hence, a software solution is needed. The package presented here is part of the [OpenRepgrid project](http://openrepgrid.org/), which contains several software packages for the analysis of repertory grid data. In the following, we will give a brief example of how a repertory grid (often abbreviated as *grid* or *repgrid*) dataset can be analyzed using the [`OpenRepGrid.ic`](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) software.  
+The [repertory grid technique](https://en.wikipedia.org/wiki/Repertory_grid) (often abbreviated as *grid* or *repgrid*) is a person-centered data collection method which generates a mix of qualitative and quantitative data. The [`OpenRepGrid.ic` R package](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) presented here is a browser-based software to perform *Interpretive Clustering* (IC) for such data. IC is a method for clustering the attributes (so called constructs) resulting from a repertory grid interview by finding relational patterns between them. The IC method is described in detail in our recent publication [[@burr_qualitative_2020]](https://doi.org/10.1080/14780887.2020.1794088). The package is part of the [OpenRepgrid project](http://openrepgrid.org/), which contains several software packages for the analysis of repertory grid data. In the following, we will give a brief example of how a repertory grid dataset can be analyzed and interpreted using the [`OpenRepGrid.ic`](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) software. 
+
+
+# Statement of need 
+
+Currently, the IC method is not implemented in any other existing repertory grid software. While IC can also be conducted by hand, this is very time consuming, error-prone and only feasible for small-sized grids. Hence, a software solution to support the IC procedure is needed. Without proper software support which facilitates the process, the IC method is likely to remain a purely methodological contribution which will rarely be used in research due to above mentioned reasons.
 
 
 # Repertory Grid Technique 
@@ -57,7 +62,7 @@ The repertory grid technique (RGT) is a method which originated from *Personal C
 # Interpretive Clustering
 
 IC is an idiographic method of interpretation which makes use of the qualitative and quantitative grid data. 
-In contrast to thematic analysis on content analysis, which could be used to identify themes in person's elicited constructs [see @braun_using_2006], IC identifies the relation between the constructs via quantitative assessment. By assessing the construct relations, implications which constructs hold for each other are identified. IC essentially identifies clusters of constructs which hold implications for each other. These clusters (or cliques) of constructs form the basis for a subsequent qualitative interpretation. We will provide a brief interpretation example below. The example builds on the analysis results the software generates as shown next. However, for a more thorough introduction to the IC method and comprehensive illustrations of interpretations, the reader is referred to our publication [[@burr_qualitative_2020]](https://doi.org/10.1080/14780887.2020.1794088). 
[…]
