### 2021-01-20 830c671
more infos for non-experts
