### 2021-05-17 #2951 @aj2duncan
Hi @gkthiruvathukal,

This looks really interesting and thank you very much for the offer but I'm not going to be able to contribute for a couple of weeks. If you're happy to wait, I'll gladly look at it then but otherwise I'll have to pass.

Thanks,

Andrew
