### 2021-05-19 #3292 @whedon
[REVIEW]: OpenRepGrid.ic: A software for Interpretive Clustering
**Submitting author:** @markheckmann (<a href="http://orcid.org/0000-0002-0736-7417">Mark Heckmann</a>)
**Repository:** https://github.com/markheckmann/OpenRepGrid.ic
**Branch with paper.md** (empty if default branch):  joss
**Version:** v0.6.1_JOSS_submission
