### 2021-05-27 #3292 @brunaw
Dear @markheckmann,

Here are a few of my initial comments/notes about the paper and code. Could you review and give any answers/solutions for what I've pointed out? After that, I'll do my review checklist for your submission:

- The paper seems to be well written in general

- I am a little bit confused about whether the focus of the paper is on the shiny app or the IC implementation as an `R` package. Is the IC implementation new and the shiny app is there for demonstration purposes only? I think that's happening because everything is being described as the 'software for IC'

- In the "Repertory Grid Technique" section, the transition between talking about the data and the results shown in the plot could be smoother (lines 36 and 37). The text seems to assume that the reader is familiar with what is usually plotted in Figure 1

- The figures have two 'Figure x' labels

- I recommend not using terms such as "is shown below" and use the figure numbers instead

- The https://ic.openprepgrid.org/ URL couldn't be reached here (as it is in the paper), but http://ic.openrepgrid.org/ works correctly

- Likewise, the docker image needs a user login. Is that the standard for all Docker users or only allowed users can see the app?

- The shiny app won't open locally to me due to functions not being correctly loaded:
`> Error in rightSidebar: could not find function "rightSidebar"`

- Not all functions are fully documented, for example:
  - `count_matches()` , `replace_one()`

I understand that those might be just helper functions, but ideally, they should be properly documented and organized like the rest of the code

- There are a few 'random' lines of comments in the `R` files (e.g. 02-calculate.R file, line 142)

Thank you,

Bruna
