### 2021-06-11 markheckmann/OpenRepGrid.ic#4 @aj2duncan
Shiny app won't load
Hi @markheckmann,

From the joss review openjournals/joss-reviews#3292 I have installed both the CRAN and github versions of the package but the shiny app won't load. The error message reported is:

```
Warning: Error in rightSidebar: could not find function "rightSidebar"
  [No stack trace available]
```

Thanks.
