### 2021-06-11 #3292 @markheckmann
@aj2duncan : Thanks, I will have a look at it. Could you add info which OS and which R version you use? Best would be a copy of  the `sessionInfo()` output. Thanks!
