### 2021-06-24 markheckmann/OpenRepGrid.ic#4 @aj2duncan
Hi @markheckmann,

Just wondering if you've had a chance to think about a timeline for this yet? We can't really move  openjournals/joss-reviews#3292 too far forward until this is resolved. All the best.
