### 2021-06-26 5099bf4
integrate breaking changes from shinyDashBoardPlus 2.0.0 (fix #4)
