### 2021-06-26 markheckmann/OpenRepGrid.ic#4 @markheckmann
Hi @aj2duncan: The `shinyDashboardPlus` package introduced significant breaking changes with version 2.0.0. This caused the issue. I fixed this now. Sorry for having been a bit slow to react! Thanks!  PS. The fixed version 0.5.1 will probably arrive on CRAN within the next few days.
