### 2021-06-26 #3292 @aj2duncan
@markheckmann, great - thanks. I'll get back to testing the app. @brunaw, has made some queries/comments☝️, have you had a chance to look at any of them?

Thanks again for fixing the app. I'll get back to you as soon as I can.
