### 2021-06-27 4cdfa29
add Dockerfile (generated by golem)
