### 2021-06-27 053b7bb
code cleanup, document helper funs
