+++ man/replace-poles.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/02-calculate.R
+\name{replace_one}
+\alias{replace_one}
+\alias{replace_all}
+\title{Replace left or right pole of constructs with blanks}
+\usage{
+replace_one(x, first = TRUE)
+
+replace_all(x, first = TRUE)
+}
+\arguments{
+\item{x}{Character vector of constructs with `@` sign used to separate construct poles.}
+
+\item{first}{Replace first pole (`TRUE`) or second pole (`FALSE`)}
+}
+\description{
+To display the two poles of the constructs as vertex labels with different 
+colors, the poles have to be printed twice. The helper function replaces all letters
+from one of the poles with blanks.
+}
+\examples{
+OpenRepGrid.ic:::replace_all("left pole @ right pole", first = TRUE)
+}
+\keyword{internal}
