### 2021-06-27 dcf755a
remove duplicated figure tag
