+++ paper.md
@@ -56,7 +56,7 @@ Currently, the IC method is not implemented in any other existing repertory grid
 The repertory grid technique (RGT) is a method which originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for clinical psychology but quickly spread to other disciplines like marketing, political, organization, and educational research in the decades after its inauguration [@fransella_manual_2004]. The RGT is a person-centered method which focuses on understanding how an individual sees, or in constructivist terms *construes*, the world. The data collected by the RGT is both, *qualitative* and *quantitative*. The qualitative part of the data consists of a list of elicited bipolar attributes (e.g. *light-hearted vs. depressed*). These are called *constructs* in PCT terminology and constitute the templates a person uses to construe a set of objects under consideration (e.g. persons like *my father*, *my mother*; so called *elements* in PCT terminology). In an additional step, quantitative data is generated by assigning a score (e.g. 1 to 6) for each element on each self-generated bipolar attribute (construct). Originally, only binary ratings (0/1) were applied, where each element was assigned to one of the construct poles. Nowadays, scales with more grades (e.g. 1 to 6) are also very common. The IC method currently works with binary ratings only. In the grid data set in Figure 1, the element *my father* received a score of 1 on the *isolated = 0 vs. sociable = 1* construct, indicating that the father is seen as sociable. The results of the RGT are usually displayed as a constructs *x* elements matrix as shown below. A more comprehensive introduction to the RGT can be found in @fransella_manual_2004.
 
 
-![**Figure 1.** Example of a binary repertory grid dataset. The data is a binarized version of a grid from @boker_reconstruction_1996](img/01-bertin-binary.png){ width=90% }
+![Example of a binary repertory grid dataset. The data is a binarized version of a grid from @boker_reconstruction_1996](img/01-bertin-binary.png){ width=90% }
 
 
 # Interpretive Clustering
@@ -76,11 +76,11 @@ ic()
 
 Alternatively, there is a web version available under https://ic.openprepgrid.org. Also, a docker image is provided under https://hub.docker.com/repository/docker/markheckmann/openrepgrid.ic. Below a screenshot of the user interface is displayed. The image shows a grid dataset from an interview with *Sylvia* which is also contained in the package.
 
-![**Figure 2.** Screenshot of the software.](img/02-screenshot.png)
+![Screenshot of the software.](img/02-screenshot.png)
 
 The results of the IC analysis are not displayed interactively but are included in an MS Excel file that can be generated and downloaded. Also the results of the intermediate IC steps as described in [Burr, King, and Heckmann (2020)](https://doi.org/10.1080/14780887.2020.1794088) are contained. The main purpose of the software is to automate the cluster identification step of the IC procedure, which is a cumbersome and error-prone task if performed manually. In Figure 3, an extract of the analysis results for Sylvia's grid and corresponding analysis settings as shown in Figure 2 are displayed and subsequently discussed.  
 
-![**Figure 3.** Results of IC method for Sylvia.](img/03-analysis-result.png){ width=90% }
+![Results of IC method for Sylvia.](img/03-analysis-result.png){ width=90% }
 
 
 # Interpretation
