+++ paper.Rmd
@@ -52,20 +52,19 @@ The [repertory grid technique](https://en.wikipedia.org/wiki/Repertory_grid) (of
 
 # Statement of need 
 
-Currently, the IC method is not implemented in any other existing repertory grid software. While IC can also be conducted by hand, this is very time consuming, error-prone and only feasible for small-sized grids. Hence, a software solution to support the IC procedure is needed. Without proper software support which facilitates the process, the IC method is likely to become a methodological contribution which will rarely be used in research due to above mentioned reasons.
+Currently, the IC method is not implemented in any other existing repertory grid software. While IC can also be conducted by hand, this is very time consuming, error-prone and only feasible for small-sized grids. Hence, a software solution to support the IC procedure is needed. Without proper software support that facilitates the IC steps, the  method is too laborious and is thus likely to become a methodological contribution which will rarely be used in research due to above mentioned reasons.
 
 
 # Repertory Grid Technique 
 
-The repertory grid technique (RGT) is a method which originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for clinical psychology but quickly spread to other disciplines like marketing, political, organization, and educational research in the decades after its inauguration [@fransella_manual_2004]. The RGT is a person-centered method which focuses on understanding how an individual sees, or in constructivist terms *construes*, the world. The data collected by the RGT is both, *qualitative* and *quantitative*. The qualitative part of the data consists of a list of elicited bipolar attributes (e.g. *light-hearted vs. depressed*). These are called *constructs* in PCT terminology and constitute the templates a person uses to construe a set of objects under consideration (e.g. persons like *my father*, *my mother*; so called *elements* in PCT terminology). In an additional step, quantitative data is generated by assigning a score (e.g. 1 to 6) for each element on each self-generated bipolar attribute (construct). Originally, only binary ratings (0/1) were applied, where each element was assigned to one of the construct poles. Nowadays, scales with more grades (e.g. 1 to 6) are also very common. The IC method currently works with binary ratings only. In the grid data set in Figure 1, the element *my father* received a score of 1 on the *isolated = 0 vs. sociable = 1* construct, indicating that the father is seen as sociable. The results of the RGT are usually displayed as a constructs *x* elements matrix as shown below. A more comprehensive introduction to the RGT can be found in @fransella_manual_2004.
-
[…]
