+++ paper.md
@@ -53,14 +53,14 @@ Currently, the IC method is not implemented in any other existing repertory grid
 
 # Repertory Grid Technique 
 
-The repertory grid technique (RGT) is a method which originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for clinical psychology but quickly spread to other disciplines like marketing, political, organizational, and educational research in the decades after its inauguration [@fransella_manual_2004]. The RGT is a person-centered method which focuses on understanding how an individual sees, or in constructivist terms *construes*, the world. The data collected by the RGT is both, *qualitative* and *quantitative*. The qualitative part of the data consists of a list of bipolar attributes (e.g. *light-hearted vs. depressed*), which are elicited during the repertory grid interview. These attributes are called *constructs* in PCT terminology. Each construct consists of two *poles*, e.g. *light-hearted* and *depressed*, with the poles usually being opposites. In PCT, the constructs constitute the templates a person uses to construe a set of objects under consideration. These objects are called *elements* in PCT terminology and are, for example, persons, brands, countries, companies etc. In our example in Figure 1, the elements are a set of persons relevant to the grid interviewee, for example, *father*, *mother*, *self*, or *ideal self*. The quantitative part of a grid consists in assigning a score (e.g. 0/1, or 1 to 6) to each element on each self-generated bipolar construct. The result of the RGT is usually displayed as a constructs *x* elements matrix with the cells containing the ratings of each element on each construct. Figure 1 shows a grid matrix where binary scores (0/1) were used to rate the elements. For example, the element *father* received a score of 1 on the *isolated = 0 vs. sociable = 1* construct, indicating that the father is construed as sociable. The *mother*, in contrast, received a score of 0 and is thus construed as isolated. The grid matrix is the central result of the RGT and yields a) a set of bipolar constructs which are relevant to the person and b) an assessment of each element (persons in Figure 1) on these dimensions. The white/red color coding in Figure 1 additionally indicates which construct pole (white = left, red = right) applies. In its original form [see @kelly_psychology_1955] and in our example, grid ratings are binary (0/1), i.e., each element was assigned to one construct pole. Nowadays, also scales with more grades (e.g. 1 to 6) are common. A general and more comprehensive introduction to the RGT can be found in @fransella_manual_2004.
[…]
