### 2021-06-27 #3292 @markheckmann
Dear @brunaw,

thanks for your comments and apologies for the late reply!  I inserted my answeres to your comments below. There is a new version of the manuscript available now in which I made the changes accordingly.

Thanks for your time and effort!

Best wishes
Mark

________

> I am a little bit confused about whether the focus of the paper is on the shiny app or the IC implementation as an R package. Is the IC implementation new and the shiny app is there for demonstration purposes only? I think that's happening because everything is being described as the 'software for IC'

The paper is on the shiny app. It is the only software which implements the IC method. The software is necessary as the steps of the IC procedure are too laborious to be done by hand. So researchers will need a software to do it.  We tried to explain this in the *statement of need* section. Please let me know if this is still too unclear and if so, could you maybe indicate which parts you find confusing?

> In the "Repertory Grid Technique" section, the transition between talking about the data and the results shown in the plot could be smoother (lines 36 and 37). The text seems to assume that the reader is familiar with what is usually plotted in Figure 1

I rearranged and slightly extended the section to make it more readable and give the reader a better understandiong what he/she sees in Figure 1. The paragraph is supposed to give readers who are unfamiliar with repertory grids a rough idea about the resulting data structure. A more thorough introduction to the RGT is, however, beyond the scope of this paper, I think.

> The figures have two 'Figure x' labels

Thanks, I corrected this.

> I recommend not using terms such as "is shown below" and use the figure numbers instead

Thanks. I changed this in the latest version of the manuscript.

> The https://ic.openprepgrid.org/ URL couldn't be reached here (as it is in the paper), but http://ic.openrepgrid.org/ works correctly

I changed the url to `http`. The app is currently hosted on a free tier shinyapps.io account, where SSL is not supported.
I plan to move it to my own server with SSL support in the future. Then both protocols (http and https) will be supported, so http seems to be the best choice for the manuscript.

> Likewise, the docker image needs a user login. Is that the standard for all Docker users or only allowed users can see the app?

The docker image is public, i.e. anyone can pull and use it without authentication.  Assuming you have the docker deamon installed and running, just type `docker pull markheckmann/openrepgrid.ic` into the console/terminal and it should pull the latest image from dockerhub. I also added instructions on how to run the image locally under the [dockerhub link](https://hub.docker.com/repository/docker/markheckmann/openrepgrid.ic) in the manuscript.

> The shiny app won't open locally to me due to functions not being correctly loaded: Error in rightSidebar: could not find function "rightSidebar"

This is the same issue @aj2duncan had and due to breaking changes in a package dep. I fixed it in [release v0.5.1](https://github.com/markheckmann/OpenRepGrid.ic/releases/tag/v0.5.1).

> Not all functions are fully documented, for example: count_matches() , replace_one() I understand that those might be just helper functions, but ideally, they should be properly documented and organized like the rest of the code

All functions, including all helpers, are documented know.

> There are a few 'random' lines of comments in the R files (e.g. 02-calculate.R file, line 142)

I cleaned up the code. There are a few commented out lines of R code which I want to keep where I added an extra comment why they are there.
