### 2021-08-20 #3292 @aj2duncan
Hi @markheckmann,

My apologies for the delay in getting back to this review. I'm happy with the paper and I think there are just a few minor things for you to address before I can complete the review.

- Can you confirm which of the [OSI approved](https://opensource.org/licenses/alphabetical) licenses you are using?
- The documentation in the repository/app is missing some of the detail that is in your paper. This includes the statement of need, functionality documentation and community guidelines. I think that the detail in the paper is fine but could be completely or partially replicated in the repo. Most of the functionality is documented in the app itself but it took me a minute to work out that "The results of the IC analysis are not displayed interactively but are included in an MS Excel file that can be generated and downloaded." as stated in the paper. It is implied by the popup in the app itself but most shiny users would probably expect the graph to be drawn in the app itself.
- The step of loading `library(OpenRepGrid.ic)` before running `ic()` is missing from the repo README. As above, it is present in the paper.

Please let me know if you've got any queries or comments about any of this.

Thanks,

Andrew
