### 2022-03-10 #3292 @markheckmann
@gkthiruvathukal: well, what can I say 🙈🙈🙈. I have found time to implement some of the changes. But especially the request to add an interactive analysis output (not only the static Excel analysis to download) will require some more time. I have two weeks of holiday in April, in which I am confident to find enough time to finish the project. So yes, more time would be appreciated.
