### 2022-04-02 markheckmann/OpenRepGrid.ic#6 @sambrhim
The number of matches between constructs* is incorrect in certain cells.
For example, in the matching matrix (below) of the following grid, the matches between C1 and C17 or C18 are given as "4" instead of "0". One can also see a diagonal of "4" which is not correct.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:////Users/samiibrahim/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip.htm">
<link rel=File-List
href="file:////Users/samiibrahim/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip_filelist.xml">
<style>

</style>
</head>

<body link="#0563C1" vlink="#954F72">

0 | E4 | E3 | E2 | E1 | 1 | preferred
-- | -- | -- | -- | -- | -- | --
C1+ | 0 | 0 | 0 | 0 | C1- | 0
C2+ | 0 | 0 | 0 | 0 | C2- | 0
C3+ | 0 | 0 | 0 | 1 | C3- | 0
C4+ | 0 | 0 | 1 | 0 | C4- | 0
C5+ | 0 | 0 | 1 | 1 | C5- | 0
C6+ | 0 | 1 | 0 | 0 | C6- | 0
C7+ | 0 | 1 | 0 | 1 | C7- | 0
C8+ | 0 | 1 | 1 | 0 | C8- | 0
C9+ | 0 | 1 | 1 | 1 | C9- | 0
C10+ | 1 | 0 | 0 | 0 | C10- | 0
C11+ | 1 | 0 | 0 | 1 | C11- | 0
C12+ | 1 | 0 | 1 | 0 | C12- | 0
C13+ | 1 | 0 | 1 | 1 | C13- | 0
C14+ | 1 | 1 | 0 | 0 | C14- | 0
C15+ | 1 | 1 | 0 | 1 | C15- | 0
C16+ | 1 | 1 | 1 | 0 | C16- | 0
C17+ | 1 | 1 | 1 | 1 | C17- | 0
C33+ | 1 | 1 | 1 | 1 | C33- | 0

</body>

</html>

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:////Users/samiibrahim/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip.htm">
<link rel=File-List
href="file:////Users/samiibrahim/Library/Group%20Containers/UBF8T346G9.Office/TemporaryItems/msohtmlclip/clip_filelist.xml">
<style>

</style>
</head>

<body link="#0563C1" vlink="#954F72">

Number   of matches between constructs* |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
  |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |   |
constructs |   | C1 | C2 | C3 | C4 | C5 | C6 | C7 | C8 | C9 | C10 | C11 | C12 | C13 | C14 | C15 | C16 | C17 | C18
C1- - C1+ | C1 |   | 4 | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 | 4 | 4
C2- - C2+ | C2 | 4 |   | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 | 4 | 4
C3- - C3+ | C3 | 3 | 3 |   | 2 | 3 | 2 | 3 | 3 | 2 | 2 | 3 | 3 | 2 | 3 | 2 | 4 | 3 | 3
C4- - C4+ | C4 | 3 | 3 | 2 |   | 3 | 2 | 3 | 3 | 2 | 2 | 3 | 3 | 2 | 3 | 4 | 2 | 3 | 3
C5- - C5+ | C5 | 2 | 2 | 3 | 3 |   | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 | 4 | 3 | 3 | 2 | 2
C6- - C6+ | C6 | 3 | 3 | 2 | 2 | 3 |   | 3 | 3 | 2 | 2 | 3 | 3 | 4 | 3 | 2 | 2 | 3 | 3
C7- - C7+ | C7 | 2 | 2 | 3 | 3 | 2 | 3 |   | 2 | 3 | 3 | 2 | 4 | 3 | 2 | 3 | 3 | 2 | 2
C8- - C8+ | C8 | 2 | 2 | 3 | 3 | 2 | 3 | 2 |   | 3 | 3 | 4 | 2 | 3 | 2 | 3 | 3 | 2 | 2
C9- - C9+ | C9 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 |   | 4 | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3
C10- - C10+ | C10 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 | 4 |   | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3
C11- - C11+ | C11 | 2 | 2 | 3 | 3 | 2 | 3 | 2 | 4 | 3 | 3 |   | 2 | 3 | 2 | 3 | 3 | 2 | 2
C12- - C12+ | C12 | 2 | 2 | 3 | 3 | 2 | 3 | 4 | 2 | 3 | 3 | 2 |   | 3 | 2 | 3 | 3 | 2 | 2
C13- - C13+ | C13 | 3 | 3 | 2 | 2 | 3 | 4 | 3 | 3 | 2 | 2 | 3 | 3 |   | 3 | 2 | 2 | 3 | 3
C14- - C14+ | C14 | 2 | 2 | 3 | 3 | 4 | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 |   | 3 | 3 | 2 | 2
C15- - C15+ | C15 | 3 | 3 | 2 | 4 | 3 | 2 | 3 | 3 | 2 | 2 | 3 | 3 | 2 | 3 |   | 2 | 3 | 3
C16- - C16+ | C16 | 3 | 3 | 4 | 2 | 3 | 2 | 3 | 3 | 2 | 2 | 3 | 3 | 2 | 3 | 2 |   | 3 | 3
C17- - C17+ | C17 | 4 | 4 | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 |   | 4
C33- - C33+ | C18 | 4 | 4 | 3 | 3 | 2 | 3 | 2 | 2 | 3 | 3 | 2 | 2 | 3 | 2 | 3 | 3 | 4 |

</body>

</html>
