### 2022-04-03 markheckmann/OpenRepGrid.ic#6 @markheckmann
Hi Sami, thanks for reporting the bug! Is that output from the resulting Excel file? Can you please provide the raw data as well? Also, what settings did you use for the analysis?
