### 2022-04-04 markheckmann/OpenRepGrid.ic#6 @sambrhim
Hallo Mark,

im Anhang findest du ein RStudio-Projekt, das alle Informationen enthält. Ein Minimalbeispiel mit library(reprex) habe ich nicht erstellen können. Für weitere Fragen stehe ich gerne zur Verfügung und würde mich freuen, mit Dir persönlich zu sprechen.

Liebe Grüße

Sami

[20220404si2mh.zip](https://github.com/markheckmann/OpenRepGrid.ic/files/8411694/20220404si2mh.zip)
