### 2022-04-19 #3292 @gkthiruvathukal
@markheckmann  Can you give me an update? I'd like to move forward on acceptance soon. Having Excel download does not sound difficult. At the same time, it does not seem "mission-critical" either. If you have acted upon all other suggestions, I think we could move forward here. Can you let me know?
