### 2022-04-24 markheckmann/OpenRepGrid.ic#6 @markheckmann
In the next release version `v0.6.0` (coming soon), there will an additional matrix that more closely corresponds to the method section and is like the one you expected. Here is a preview of the output.

![image](https://user-images.githubusercontent.com/152927/164977762-56418235-7a6d-4eb7-b325-57e4c86b5412.png)
