### 2022-04-24 markheckmann/OpenRepGrid.ic#6 @markheckmann
One hint: you can create the Excel file output using your file without using the shiny UI as follows (see package docs `?OpenRepGrid.ic`).
```R
library(tidyverse)
library(openxlsx)
library(igraph)
library(OpenRepGrid.ic)

file <-"kgw.xlsx"
x <- read.xlsx(file)
tests <- check_excel_input(x)
l <- network_graph_images(x, min_clique_size = 3, show_edges = TRUE, min_matches = 3)
file_tmp <- create_excel_output(file, l)
file.show(file_tmp)
```
