### 2022-04-24 357f9dd
New Excel output (close #6), update README, tweak UI (e.g. mouse-overs)
* Excel output: added table with raw matching score, i.e. without constructs
  reversal, to better resemble step 3 from method description
* Update README with parts from the JOSS paper as suggested by reviewers
* UI Tweaks to improve usability: info box informs that the output is an
  Excel file, mouser-over for most analysis settings
* Added MIT License
