### 2022-04-24 912dcb6
add R check standard github actions YAML
