+++ .github/workflows/R-CMD-check.yaml
@@ -3,6 +3,8 @@
 on:
   push:
     branches: [main, master]
+    paths-ignore:
+      - 'README.md'
   pull_request:
     branches: [main, master]
 
