### 2022-04-24 2df101d
move from Travis / Appveyor to GH Actions [no ci]
