### 2022-05-08 b5def85
grid KPIs boxes shown again after loading data
