### 2022-05-08 761d019
change README as suggested by JOSS reviewers
Following JOSS reviwers suggestion that the `detail in
the paper is fine but could be completely or partially
replicated in the repo`. The paper is now replicated in
the README with only slight modifications.
