+++ README.md
@@ -1,76 +1,239 @@
-# Interpretive Clustering
-
 <!-- badges: start -->
+
 [![](https://www.r-pkg.org/badges/version/OpenRepGrid.ic?color=success)](https://cran.r-project.org/package=OpenRepGrid.ic)
-[![Codecov test coverage](https://codecov.io/gh/markheckmann/OpenRepGrid.ic/branch/master/graph/badge.svg)](https://codecov.io/gh/markheckmann/OpenRepGrid.ic?branch=master)
+[![Codecov test
+coverage](https://codecov.io/gh/markheckmann/OpenRepGrid.ic/branch/master/graph/badge.svg)](https://codecov.io/gh/markheckmann/OpenRepGrid.ic?branch=master)
 [![R-CMD-check](https://github.com/markheckmann/OpenRepGrid.ic/workflows/R-CMD-check/badge.svg)](https://github.com/markheckmann/OpenRepGrid.ic/actions)
 <!-- badges: end -->
 
+## Introduction
 
-This package implements the [Interpretive Clustering (IC)](https://doi.org/10.1080/14780887.2020.1794088) 
-method as described in [Burr, King, and Heckmann (2020)](https://doi.org/10.1080/14780887.2020.1794088).
-An introduction to the software is available on [YouTube](https://youtu.be/f9oINYA22Rc).
-Interpretive Clustering is a variant of construct clustering for [repertory grid](https://en.wikipedia.org/wiki/Repertory_grid)
-data. While derived from theoretical considerations based on [Personal Construct Theory](https://en.wikipedia.org/wiki/Personal_construct_theory), the procedure itself is mathematically equivalent to a problem from graph theory called [maximal cliques enumeration](https://en.wikipedia.org/wiki/Clique_problem#Listing_all_maximal_cliques). 
-Given a similarity measure, in our case the number of matching scores between
-two constructs, a network graph of relatedness between constructs is construed.
-A clique is a group of constructs which are all mutually related, given some
-cut-off criterion for relatedness (e.g. 6 matching scores in a grid with 7
-elements). While an offline approach is also described to find the construct
-cliques, this software automates the process. Below you see the resulting
-cliques for Sylvia's sample grid. In the paper, the interpretation of the
-discovered cliques is discussed in detail.
-                       
-![example](inst/shiny/www/sylvia_cliques.png "Construct cliques for Sylvias's grid")
+The [`OpenRepGrid.ic` R
+package](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html)
+is a browser-based software to perform *Interpretive Clustering* (IC)
+for repertory grid data. The [repertory
+grid](https://en.wikipedia.org/wiki/Repertory_grid) (often abbreviated
+as *grid* or *repgrid*) is a person-centered data collection method. It
+is primarily used in clinical psychology and generates a mix of
+qualitative and quantitative data. IC is a method for clustering the
+attributes (so called constructs) resulting from a repertory grid
+interview by finding relational patterns between them. The IC method is
+described in more detail in our publication [(Burr, King, and Heckmann
+2020)](https://doi.org/10.1080/14780887.2020.1794088). An introduction
+to the software is also available on
+[YouTube](https://youtu.be/f9oINYA22Rc). The package is part of the
+[OpenRepgrid project](http://openrepgrid.org/), which contains several
+software packages for the analysis of repertory grid data. In the
+following, we will briefly explain, why the software was developed,
+describe the repertory grid technique and provide a short example of how
+a grid dataset can be analyzed and interpreted using our
+[`OpenRepGrid.ic`](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html)
+software.
 
+## Statement of need
 
-## Statement of need 
+Currently, the IC method is not implemented in any other existing
[…]
