+++ README.Rmd
@@ -1,0 +1,74 @@
+---
+title: "Interpretive Clustering"
+bibliography: inst/extdata/bibliography.bib
+output:
+  html_document:
+    keep_md: yes
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+<!-- badges: start -->
+[![](https://www.r-pkg.org/badges/version/OpenRepGrid.ic?color=success)](https://cran.r-project.org/package=OpenRepGrid.ic)
+[![Codecov test coverage](https://codecov.io/gh/markheckmann/OpenRepGrid.ic/branch/master/graph/badge.svg)](https://codecov.io/gh/markheckmann/OpenRepGrid.ic?branch=master)
+[![R-CMD-check](https://github.com/markheckmann/OpenRepGrid.ic/workflows/R-CMD-check/badge.svg)](https://github.com/markheckmann/OpenRepGrid.ic/actions)
+<!-- badges: end -->
+
+## Introduction
+
+The [`OpenRepGrid.ic` R package](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) is a browser-based software to perform *Interpretive Clustering* (IC) for repertory grid data. The [repertory grid](https://en.wikipedia.org/wiki/Repertory_grid) (often abbreviated as *grid* or *repgrid*) is a person-centered data collection method. It is primarily used in clinical psychology and generates a mix of qualitative and quantitative data. IC is a method for clustering the attributes (so called constructs) resulting from a repertory grid interview by finding relational patterns between them. The IC method is described in more detail in our publication [[@burr_qualitative_2020]](https://doi.org/10.1080/14780887.2020.1794088). An introduction to the software is also available on [YouTube](https://youtu.be/f9oINYA22Rc). The package is part of the [OpenRepgrid project](http://openrepgrid.org/), which contains several software packages for the analysis of repertory grid data. In the following, we will briefly explain, why the software was developed, describe the repertory grid technique and provide a short example of how a grid dataset can be analyzed and interpreted using our [`OpenRepGrid.ic`](https://cran.r-project.org/web/packages/OpenRepGrid.ic/index.html) software. 
+
+
+## Statement of need 
+
+Currently, the IC method is not implemented in any other existing repertory grid software. While IC can also be conducted by hand, this is very time consuming, error-prone and only feasible for small-sized grids. Hence, a software solution to support the IC procedure is needed. Without proper software support that facilitates the IC steps, the  method is too laborious and is thus likely to become a methodological contribution which will rarely be used in research due to above mentioned reasons.
+
+
+## Repertory Grid Technique 
+
[…]
