+++ paper.md
@@ -20,7 +20,7 @@ affiliations:
   index: 1
 - name: University of Huddersfield, GB
   index: 2
-date: 10 January 2021
+date: 08 May 2022
 bibliography: paper.bib
 output:
   html_document:
