### 2022-05-08 #3292 @markheckmann
Dear all,

thanks for your patience and my apologies with regard to the timeliness of our revisions.

@aj2duncan: please find below our answers to your comments:

> Can you confirm which of the [OSI approved](https://opensource.org/licenses/alphabetical) licenses you are using?

The software uses the MIT license. The license statement is now contained in the package’s DESCRIPTION file. A copy of the MIT license itself is additionally found in the root folder of the  R package repo.

> The documentation in the repository/app is missing some of the detail that is in your paper. This includes the statement of need, functionality documentation and community guidelines. I think that the detail in the paper is fine but could be completely or partially replicated in the repo.

As you suggested, we now replicated the text from the paper (with minor modifications only) directly in the README of the repo.

> Most of the functionality is documented in the app itself but it took me a minute to work out that "The results of the IC analysis are not displayed interactively but are included in an MS Excel file that can be generated and downloaded." as stated in the paper. It is implied by the popup in the app itself but most shiny users would probably expect the graph to be drawn in the app itself.

We now added a new info box at the top of the analysis tab of the software explaining that the final result of the analysis is a downloadable file, so the user will directly see it. We also discussed adding interactivity directly in the tool, but came to the conclusion, that the benefits would be minor. After the quantitative analysis, a researcher must thoroughly look at the results first and make a qualitative interpretation. This process takes quite some time and iterations are rather slow, so we decided against the interactive display of results.

> The step of loading library(OpenRepGrid.ic) before running ic() is missing from the repo README. As above, it is present in the paper.

We now added the line to the repo README.md

@brunaw: We already answered your comments above.

**Further changes compared to the last version**:

* We fixed an issue reported by a user (https://github.com/markheckmann/OpenRepGrid.ic/issues/6) with regard to the output of the intermediate steps of the analysis.
* We added mouse-over pop-ups to the majority of buttons, checkboxes, etc. for a smoother user experience

![image](https://user-images.githubusercontent.com/152927/167312516-db62704e-a4c7-41ec-827e-af06221f6279.png)

Please let me know if you have any further comments / see need for further changes or if we are ready to move ahead. The package version 0.6 will hit CRAN in a few days if you feel comfortable with the current status.

Thanks again!
Mark
