### 2022-05-10 4e09965
remove URL to hosted software instance
We cannot make sure that the instance will be
up and running all time, so we prefer to remove
the URL to it.
