### 2022-05-14 markheckmann/OpenRepGrid.ic#6 @sambrhim
Lieber Mark,

ich bin erst jetzt dazu gekommen, Dir zu antworten. Vielen Dank für Deine ausführliche Erklärung und den Hinweis, wie ich die Excel-Dateiausgabe mit R-Syntax erstellen kann.

Liebe Grüße

Sami

PS: Übrigens die Art und Weise, wie Du es erklärst, ist ganz nach meinem Geschmack :)

> Am 24.04.2022 um 14:09 schrieb Mark Heckmann ***@***.***>:
>
>
> Hi Sami,
>
> I had a look at the results and your remarks.
>
> The results are correct, however
> maybe slightly misleading in the way they are presented.
> 1. Example: constructs C1 vs. C18
>
>  <https://user-images.githubusercontent.com/152927/164974901-d5e775b9-1e97-45a1-b826-2acb50a0c3ea.png>
> The constructs C1 and C18 have no matches the way they are oriented. However, if one of the constructs is reversed, they have 4 matches and this a perfect inverse or negative relation. In the Excel sheet calculations, the result matrix named Number of matches between constructs counts the number of matches in either way, i.e. also with one of the constructs reversed, as constructs could be negatively related. In other words, it orients the constructs in a way, to maximize their relatedness. To indicate this, there is a negative entry at the corresponding position in the 2nd result matrix Relatedness by criterion (matches >= 3). In other words: The constructs C1 vs. C18 have 4 matches and a perfect negative relation, when on construct is reversed.
>
>  <https://user-images.githubusercontent.com/152927/164975662-b6c3f3cb-63c8-40db-8ebe-6ab1e5d3835a.png>
> 2) Presentation of results in Excel file
>
> I concede that this is not obvious at first glance. I will add an explanation inside the Excel ouput to better explain this.
> There is, however, an explanation in the method section/tab of the software:
>
> It is important to recognise that constructs may be negatively related, such that the preferred pole of one construct is related to the non-preferred pole of another. It is important to capture these negative relationships as well as the positive ones, since these can indication tensions and contradictions within the person’s construing. So in inspecting the construct pairs you should also identify any pair where there were instances of this (ie a tick against an element for one construct and a cross for the other). In the case of our research, we looked for cases where there were 5 or more mis-matches, ie. 2 or fewer ‘matches’.
>
> So basically, the matrix as shown in the output reflects the way the analysis has been implented technically. Instead of using a low and high number of matches to indicate relatedness, in my implementation the maximal number of matches under optional reversal is used. The result is the same, but the implementation was easier for me, as I recall. Yet, I agree that this deviates from the description in the methods section. I will add this as a future improvement issue.
>
> HTH
> Mark
>
