### 2022-07-17 #3292 @brunaw
Thanks all, I've finished my review now, my apologies for the delay. I only have a few quick suggestions/comments:

- Because this is a shiny app, I think it would be nice to have a gif in the package README file showing how to use the app instead of just a screenshot of the interface, or at least a few more screenshots indicating the steps to run the analysis
- Also on the exemplification, it woul be nice to have an 'example analysis' in the shiny app (as another tab), because currently, we can only see any results if we input our own data, which might be a bit confusing for someone who is exploring the app and maybe doesn't have their own data yet
- I found some comments in the code that are in German. I think it would be better to at least have them translated to english, so when people are trying to understand what the functions do or contribute to the work, they won't have this language barrier
- The tests seemed a bit incomplete to me, it doesn't look like all functions are being tested, just the app function and the file format. Is this enough to test the whole code?
