### 2022-08-27 df0a04d
All code comments in English now
