+++ R/01-utils.R
@@ -66,7 +66,7 @@ dt_default <- function(text = "Waiting for data ...",
 #' @examples
 #' 
 #' cell_text_split("10, 20,30" )
-#' cell_text_split("; ,  10  ,,,  20;30,," ) # sehr robust
+#' cell_text_split("; ,  10  ,,,  20;30,," ) # very robust
 #' 
 cell_text_split <- function(x)
 {
@@ -83,11 +83,11 @@ cell_text_split <- function(x)
 make_names_vec <- function(x) 
 {
   nms <- tolower(x)
-  nms <- stringr::str_replace_all(nms, "[[:blank:]]+", "_")       # Leerzeichen ersetzen
-  nms <- stringr::str_replace_all(nms, "\\.|-|/|\\(|\\)|&|\\?", "_")  # punkte, bindestrich, (back-)slashes, Fragezeichen zu unterstrich
-  nms <- stringr::str_replace_all(nms, "[_]+", "_")               # mehrere unterstriche zu einem
-  nms <- stringr::str_replace_all(nms, "[_]+$", "")               # unterstriche am ende entfernen
-  nms <- stringr::str_replace_all(nms, "\u00E4", "ae")            # umlaute ersetzen
+  nms <- stringr::str_replace_all(nms, "[[:blank:]]+", "_")       # replace blanks
+  nms <- stringr::str_replace_all(nms, "\\.|-|/|\\(|\\)|&|\\?", "_")  # replace . - \ ? to _ (underscore)
+  nms <- stringr::str_replace_all(nms, "[_]+", "_")               # replace multiple underscores by one
+  nms <- stringr::str_replace_all(nms, "[_]+$", "")               # remove trailing underscores
+  nms <- stringr::str_replace_all(nms, "\u00E4", "ae")            # replace German umlauts by their two letter ASCII version
   nms <- stringr::str_replace_all(nms, "\u00FC", "ue") 
   nms <- stringr::str_replace_all(nms, "\u00F6", "oe") 
   nms
@@ -120,12 +120,3 @@ fnum <- function(x, digits = 2)
   x <- round(x, digits)
   formatC(x, digits = digits, format = "f")
 }
-
-
-
-
-
-
-
-
-
