### 2022-08-27 9b9287a
Add sample data button & fix icons
* Button to load sample data and get started right away
  without uploading any
* Update icon names using Font Awesome 6
