+++ inst/shiny/server.R
@@ -1,6 +1,6 @@
 #///////////////////////////////////////////////////////////////////////////////////////////
 #
-#                                   server
+#                                   SERVER
 #
 #//////////////////////////////////////////////////////////////////////////////////////////
 
@@ -51,21 +51,37 @@ server <- function(input, output, session)
   #### _______________________ ####
   #### UPLOAD ####
   
-  # upload grid
+  # __ UPLOAD FILE ----
+  
   observeEvent(input$excel_input, 
   {
      req(input$excel_input)
      in_file <- input$excel_input
      if (is.null(in_file))
        return(NULL)
+     rv$file_path <- in_file$datapath   # needed as load_sample_data does not populate input$excel_input
+     rv$file_name <- in_file$name
      rv$data <- openxlsx::read.xlsx(in_file$datapath)
      hide("down_btn")
      disable("btn_download_excel")
-    
      # rv$data_status <- NULL  # set back data status
      rv$number_of_uploads <- rv$number_of_uploads + 1
-   })
+  })
 
+  
+  # __ LOAD SAMPLE ----
+  
+  observeEvent(input$load_sample_data, 
+  {
+    path <- system.file("extdata", "sylvia.xlsx", package = "OpenRepGrid.ic")
+    rv$file_path <- path 
+    rv$file_name <- basename(path)
+    rv$data <- openxlsx::read.xlsx(path)
+    hide("down_btn")
+    disable("btn_download_excel")
+    rv$number_of_uploads <- rv$number_of_uploads + 1
+  })
+  
   
   #### .                       ####
   #### _______________________ ####
@@ -227,12 +243,12 @@ server <- function(input, output, session)
     
     dt <- DT::datatable(x, rownames = FALSE, colnames = nms, 
                         options = list(
-                          headerCallback = header_callback, #JS(headerCallback),
+                          headerCallback = header_callback, # JS(headerCallback),
                           paging = FALSE,
                           ordering = FALSE,
                           dom = 't',
                           columnDefs = column_defs,
-                          initComplete = DT::JS(             # change column header size along with cells
+                          initComplete = DT::JS(            # change column header size along with cells
                             paste0("function(settings, json) {",
                             "$(this.api().table().header()).css({'font-size': '", grid_font_size, "pt'});",
                             "}"))
@@ -291,10 +307,11 @@ server <- function(input, output, session)
   
   observeEvent(input$btn_process, 
   {
-    req(input$excel_input$datapath)
+    # req(input$excel_input$datapath)
+    req(rv$data)
     disable("btn_download_excel")
-    file <- input$excel_input$datapath
-
+    # file <- input$excel_input$datapath
+    file <- rv$file_path
     min_matches <- input$par_min_match
     min_clique_size <- input$par_min_clique_size
     align_poles <- input$par_align_poles
@@ -337,7 +354,8 @@ server <- function(input, output, session)
   output$btn_download_excel <- downloadHandler(
     
     filename = function() {
-      filename <- input$excel_input$name
+      # filename <- input$excel_input$name
+      filename <- rv$file_name
       str_replace(filename, ".xlsx", "_CLIQUES.xlsx")
     },
     content = function(file) {
