### 2022-08-28 0354534
Add tests raising test coverage to over 90%
