### 2022-08-28 6fdca52
Add `Sample output` tab
* New `Sample output`tab added with extracts
  from the output for Sylvia's grid
