+++ NEWS.Rmd
@@ -9,13 +9,9 @@ output: html_document
     showing how to use the app instead of just a screenshot of the interface, or at least a few more
     screenshots indicating the steps to run the analysis
 
--   Also on the exemplification, it woul be nice to have an 'example analysis' in the shiny app (as
-    another tab), because currently, we can only see any results if we input our own data, which
-    might be a bit confusing for someone who is exploring the app and maybe doesn't have their own
-    data yet
-
 ## v0.6.1
 
+-   New `Sample output`tab added with extracts from the output for Sylvia's grid
 -   Additional test raising test coverage to over 90%
 -   Button to load sample data and get started right away without uploading any data
 -   Update icon names using Font Awesome 6
