### 2022-08-28 7131ecc
Update animated software overview to README
* Add animated gif image with short software overview to
  README
