+++ paper.Rmd
@@ -89,7 +89,7 @@ The results of the IC analysis are not displayed interactively but are included 
 
 Psychologically relevant information can be obtained from the interpretation of the network graphs. What follows is a shortened example. More comprehensive examples are outlined in our publication.
 
-In the resulting diagram for Sylvia's grid in Figure 3, a construct is indicated by a circle, with (-) denoting the preferred and (-) the non-preferred pole. The diagram shows three clusters (also called cliques), indicated by the colored hulls around several constructs. In Sylvia's case, the three clusters are highly overlapping. 
+In the resulting diagram for Sylvia's grid in Figure 3, a construct is indicated by a circle, with (+) denoting the preferred and (-) the non-preferred pole. The diagram shows three clusters (also called cliques), indicated by the colored hulls around several constructs. In Sylvia's case, the three clusters are highly overlapping. 
 Two of these are of particular interest, sharing a ‘core’ of three constructs – ‘(+) Wild, free *vs* (-) controlled, contained’, ‘(+) Massive sense of space, expansive  *vs* (-) railed-off, small world’ and ‘(+) Freedom, wildness *vs*  (-) conventional’, with (+) indicating the preferred and (-) the non-preferred pole. In one cluster, these three constructs are strongly associated with ‘(+) Verdant *vs* (-) dead, nothing thriving’; the association between her preferred poles suggests that she is drawn to places that are thriving and green, wild and expansive, as opposed to those which lack life, are small-scale, controlled and conventional. However, these three constructs share another cluster with the construct ‘(+) Cosy *vs* (-) Chaotic’, where ‘cosy’ is her preferred pole. In this cluster, however, her desires for the wild, free and expansive appear to be in tension with her desire for the ‘cosy’, as they are aligned with her non-preferred pole ‘chaotic‘. The attraction of wild, free spaces for Sylvia is therefore not straightforward. 
 
 The third cluster includes the ‘(+) Verdant *vs* (-) dead, nothing thriving’ construct, which is here associated with ‘(+) Exciting, a lot going on *vs* (-) flat, unvarying, depressed, unenergetic’, ‘(+) Dramatic *vs* (-) unvarying, goes on and on’ and ‘(+) Variable *vs* (-) doesn’t change’. This suggests that to Sylvia ‘verdant’ spaces are also full of excitement, drama and variability- they are full of life in these ways. However, the fact that these three constructs do not cluster with the wild/expansive/freedom constructs indicates that they constitute a somewhat separate idea for her. A ‘wild’ space for her need not be ‘exciting’, for example, although a ‘verdant’ space is likely to be both exciting and wild. Interpretive clustering therefore gives us insight into some of the complexity of Sylvia’s construing.
