### 2022-08-28 bf66da3
fix warning in checks by removing non ASCII
