### 2022-08-28 b6d898a
make summaries work when pressing button (fix #7)
