+++ DESCRIPTION
@@ -48,14 +48,14 @@ Imports:
     scales,
     splines,
     igraph,
-    testthat,
     tidyr,
     withr,
     RColorBrewer
 Suggests:
     knitr,
     rmarkdown,
-    covr
+    covr,
+    testthat
 Encoding: UTF-8
 URL: https://github.com/markheckmann/OpenRepGrid.ic
 BugReports: https://github.com/markheckmann/OpenRepGrid.ic/issues
