### 2022-08-28 #3292 @markheckmann
Hi @brunaw, @gkthiruvathukal,

I updated the package to `v0.6.1` (not yet on CRAN) making the following changes:

* added an animated gif to the README, that shows how the software works
* all German comments are translated to English now
* no need to bring you own data anymore.  On the `Analysis` tab, you can now  simply press the `Load Sylvias's sample grid` button to load a sample dataset tab. The data can also be downbloaded as an Excel file (this is not new) to be used as a template so store your own grid data.
* there is a new tab `Sample output` showing extracts of the ouput fpr Sylvia's dataset
* Several additional tests raised the test coverage to 93% (from 56%)

I will upload the package to CRAN if there are no further change requirements from your side. Please let me know.
Thanks!
