### 2022-10-25 9187ec8
add missing imports & cosmetic tweaks
* import packages used only in ui.R and server.R
* change import directives so that no functions are masked
  thus avoiding import warnings
* Code now compliant with tidyverse style guide using
  the `styler` package
