+++ inst/shiny/ui.R
@@ -26,7 +26,6 @@ suppressWarnings({
     library(reshape2)
     library(testthat)
     library(scales)
-    library(tidyverse)
     library(glue)
     library(splines)
     library(igraph)
