### 2022-10-25 222b04f
update CRAN comments for release
