### 2022-10-25 #3292 @markheckmann
FYI: Just pushed the latest package version `v0.6.1` to CRAN: https://cran.rstudio.com/web/packages/OpenRepGrid.ic/index.html. Please, let me know in case you need anything else to proceed.
Thanks!
