### 2023-02-20 d7ad78f
add CITATION.cff (Zenodo), update Rbuildignore
