+++ README.md
@@ -2,6 +2,7 @@
 [![Codecov test
 coverage](https://codecov.io/gh/markheckmann/OpenRepGrid.ic/branch/master/graph/badge.svg)](https://codecov.io/gh/markheckmann/OpenRepGrid.ic?branch=master)
 [![R-CMD-check](https://github.com/markheckmann/OpenRepGrid.ic/workflows/R-CMD-check/badge.svg)](https://github.com/markheckmann/OpenRepGrid.ic/actions)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7655618.svg)](https://doi.org/10.5281/zenodo.7655618)
 
 # Interpretive Clustering
 
