### 2023-02-20 #3292 @markheckmann
Thanks for jumping in, @arfon . As requested, I prepared the following:

- New GitHub release tag: [v0.6.1_JOSS_submission](https://github.com/markheckmann/OpenRepGrid.ic/releases/tag/v0.6.1_JOSS_submission)
- Zenodo archive: <https://zenodo.org/record/7655618>
- Metadata was checked.
- DOI: 10.5281/zenodo.7655618

Thanks!
