### 2023-02-20 45e2dcf
Update gh-actions to v2 and add README badge
