### 2023-02-20 #3292 @arfon
@tarleb – any chance you could help with this error above from the XML validator? I'm not sure what the issue is here. Also, is there a way to see the intermediate files associated with the GitHub Action workflow @xuanxu ?

@markheckmann – while we wait for a couple of issues to be addressed here, did you check if any of those DOIs suggested above are correct for your references?
