### 2023-02-20 #3292 @xuanxu
The metadata files can be found in the generated [pull request](https://github.com/openjournals/joss-papers/pull/3984).

The error points to lines [104](https://github.com/openjournals/joss-papers/pull/3984/files#diff-51b7cc59f49cb5eb1a03af594c9a97b88be2e6ee70ddd02eeef166216ff71590R104) and [201](https://github.com/openjournals/joss-papers/pull/3984/files#diff-51b7cc59f49cb5eb1a03af594c9a97b88be2e6ee70ddd02eeef166216ff71590R201) of the JATS metadata file.
