### 2023-02-20 #3292 @tarleb
Thanks @xuanxu, I second that and the PR should fix it.

Background: References are automatically linked to the respective entry in the bibliography. Placing a reference within a link thus leads to nested links, which is not allowed in most formats. I wonder if this should be fixed in pandoc, or whether we should add a filter to check for this case.
