### 2023-02-20 612029c
Add missing DOIs to .bib file
