+++ paper.bib
@@ -1,352 +1,4 @@
 
-@article{dorough_implicative_2007,
-	title = {Implicative dilemmas and psychological well-being},
-	volume = {4},
-	journal = {Personal Construct Theory \& Practice},
-	author = {Dorough, Stefanie and Grice, James W. and Parker, Jessica},
-	year = {2007},
-	keywords = {conflict, implicative dilemma, index, repgrid},
-	pages = {83--101}
-}
-
-@techreport{bell_gridstat_2009,
-	address = {University of Melbourne, Australia},
-	type = {manual},
-	title = {Gridstat version 5 - {A} program for analyzing the data of a repertory grid},
-	institution = {Department of Psychology},
-	author = {Bell, Richard C.},
-	year = {2009},
-	keywords = {repgrid, software}
-}
-
-@article{hartmann_element_1992,
-	title = {Element comparisons in repertory grid technique: {Results} and consequences of a monte carlo study},
-	volume = {5},
-	issn = {0893-603X},
-	shorttitle = {Element {Comparisons} {In} {Repertory} {Grid} {Technique}},
-	url = {http://www.informaworld.com/10.1080/08936039208404940},
-	doi = {10.1080/08936039208404940},
-	number = {1},
-	urldate = {2010-07-25},
-	journal = {International Journal of Personal Construct Psychology},
-	author = {Hartmann, A.},
-	year = {1992},
-	keywords = {distance, index, monte carlo, repgrid, stats},
-	pages = {41--56}
-}
-
-@article{winter_construct_1982,
-	title = {Construct relationships, psychological disorder and therapeutic change},
-	volume = {55 (Pt 3)},
-	issn = {0007-1129},
-	url = {http://www.ncbi.nlm.nih.gov/pubmed/7126491},
-	abstract = {The problem of interpretation of construct relationships of abnormal content is considered in the context of the use of the repertory grid to assess therapeutic outcome in neurotics. With a view to providing a rationale for such interpretation, examples are presented of construct relationships indicative of psychopathology, and of predicted changes in them during therapy. Implications for the planning and evaluation of therapy are discussed, and the importance of exploring the personal meaning of the patient's complaints is emphasized.},
-	urldate = {2011-05-15},
-	journal = {The British Journal of Medical Psychology},
-	author = {Winter, David A.},
-	month = sep,
-	year = {1982},
-	pmid = {7126491},
-	keywords = {conflict, implicative dilemma, repgrid},
-	pages = {257--269}
-}
-
-@book{kelly_psychology_1955,
-	address = {New York},
-	title = {The psychology of personal constructs},
-	publisher = {Norton},
-	author = {Kelly, George Alexander},
-	year = {1955},
-	keywords = {PCP}
-}
-
-@misc{fromm_gridsuite_2006,
-	address = {Stuttgart},
-	title = {{GridSuite} 4},
-	abstract = {http://www.gridsuite.de},
-	publisher = {TGZ constructiv},
-	author = {Fromm, Martin and Bacher, Andreas},
-	year = {2006},
-	keywords = {repgrid, software}
-}
-
-@misc{feixas_gridcor:_2002,
-	address = {Barcelona},
-	title = {{GRIDCOR}: {Correspondence} {Analysis} for {Grid} {Data} (version 4.0) [{Computer} software and repertory grid manual]},
-	url = {www.terapiacognitiva.net/record.},
-	publisher = {Centro de Terapia Cognitiva},
-	author = {Feixas, Guillem and Cornejo, Josi Manuel},
-	year = {2002},
-	keywords = {repgrid, software}
-}
-
-@book{greenacre_biplots_2010,
-	address = {Madrid},
-	title = {Biplots in practice},
-	url = {http://www.multivariatestatistics.org/biplots.html},
-	publisher = {BBVA Foundation},
[…]
