### 2023-02-20 #3292 @markheckmann
@xuanxu Thanks for the fix! :)

@arfon, @xuanxu : I checked the suggested DOIs and added correct ones to the the .bib file.

I also checked the links in the new PDF and it looks good.

Thanks!
