### 2023-02-21 #3292 @arfon
> The metadata files can be found in the generated https://github.com/openjournals/joss-papers/pull/3984.

Thanks @xuanxu – I'm afraid I hadn't realized the PR was still opened even when these errors happen...
