### 2023-02-21 #3292 @markheckmann
> One more minor typo to fix @markheckmann: [markheckmann/OpenRepGrid.ic#10](https://github.com/markheckmann/OpenRepGrid.ic/pull/10)

@arfon : Thanks, PR is merged now
